requirements.txt
streamlit
openai
requests
beautifulsoup4
lxml
streamlit-option-menu
pyarrow
pillow
google-generativeai